Converts all kernels to double precision (needs work on accuracy) Addds _2 API calls which take separate lon/lat x/y vectors as input/output.  These calls are surprisingly a lot slower than the single interspersed vector API. Compiles all kernel code into C string constants, so they reside in the library. Adds JNI interface.

+++ src/projcl_run.h
@@ -7,10 +7,10 @@
  *
  */
 
-cl_int pl_read_buffer(cl_command_queue queue, cl_mem xy_out_buf, float *xy_out, size_t out_count);
+cl_int pl_read_buffer(cl_command_queue queue, cl_mem xy_out_buf, double *xy_out, size_t out_count);
 cl_kernel pl_find_projection_kernel(PLContext *pl_ctx, PLProjection proj, int fwd, PLSpheroid ell);
 
-cl_int pl_read_buffer_2(cl_command_queue queue, cl_mem xy_out_buf, float *x_out, float *y_out, size_t out_count);
+cl_int pl_read_buffer_2(cl_command_queue queue, cl_mem xy_out_buf, double *x_out, double *y_out, size_t out_count);
 
 cl_int pl_enqueue_projection_kernel_points(PLContext *pl_ctx, cl_kernel kernel,
         PLProjection proj, PLProjectionParams *params, PLProjectionBuffer *pl_buf);
@@ -18,15 +18,15 @@ cl_int pl_enqueue_projection_kernel_grid(PLContext *pl_ctx, cl_kernel kernel,
         PLProjection proj, PLProjectionParams *params, PLPointGridBuffer *src, PLPointGridBuffer *dst);
 
 cl_int pl_run_kernel_forward_geodesic_fixed_distance(PLContext *pl_ctx, cl_kernel kernel, 
-    PLForwardGeodesicFixedDistanceBuffer *pl_buf, float *xy_out, PLSpheroid pl_ell, double distance);
+    PLForwardGeodesicFixedDistanceBuffer *pl_buf, double *xy_out, PLSpheroid pl_ell, double distance);
 cl_int pl_run_kernel_forward_geodesic_fixed_angle(PLContext *pl_ctx, cl_kernel kernel, 
-    PLForwardGeodesicFixedAngleBuffer *pl_buf, double xy_in[2], float *xy_out, PLSpheroid pl_ell, double angle);
+    PLForwardGeodesicFixedAngleBuffer *pl_buf, double xy_in[2], double *xy_out, PLSpheroid pl_ell, double angle);
 cl_int pl_run_kernel_inverse_geodesic(PLContext *pl_ctx, cl_kernel inv_kernel,
-        PLInverseGeodesicBuffer *pl_buf, float *dist_out, PLSpheroid pl_ell, double scale);
+        PLInverseGeodesicBuffer *pl_buf, double *dist_out, PLSpheroid pl_ell, double scale);
 
 cl_int pl_run_kernel_geodesic_to_cartesian(PLContext *pl_ctx, cl_kernel g2c_kernel, 
         PLDatumShiftBuffer *pl_buf, PLSpheroid pl_ell);
 cl_int pl_run_kernel_transform_cartesian(PLContext *pl_ctx, cl_kernel transform_kernel, 
         PLDatumShiftBuffer *pl_buf, PLDatum src_datum, PLDatum dst_datum);
 cl_int pl_run_kernel_cartesian_to_geodesic(PLContext *pl_ctx, cl_kernel c2g_kernel, 
-        PLDatumShiftBuffer *pl_buf, float *xy_out, PLSpheroid pl_ell);
+        PLDatumShiftBuffer *pl_buf, double *xy_out, PLSpheroid pl_ell);
