+++ src/projcl_util.c
@@ -12,7 +12,7 @@
 #include <strings.h>
 #include <stdio.h>
 
-void _pl_copy_pad(float *dest, size_t dest_count, const float *src, size_t src_count) {
+void _pl_copy_pad(double *dest, size_t dest_count, const double *src, size_t src_count) {
 	int i;
 	for (i=0; i<src_count; i++) {
 		dest[i] = src[i];
